Функция появления и разнос по файлам
